requirements.txt
google-generativeai
requests
beautifulsoup4
python-docx
pypandoc
python-dotenv
PyPDF2
feedparser